Fix benchmarck and sandbox test with cuda

+++ keops/sandbox/test_logsumexp.cu
@@ -64,7 +64,7 @@ int main() {
 
     std::cout << "Testing LogSumExp reduction" << std::endl;
 
-    int Nx=5000, Ny=2000;
+    unsigned int Nx=5000, Ny=2000;
         
     std::vector<__TYPE__> vf(Nx*FUNCONVF::DIM);    fillrandom(vf); __TYPE__ *f = vf.data();
     std::vector<__TYPE__> vx(Nx*DIMPOINT);    fillrandom(vx); __TYPE__ *x = vx.data();
@@ -87,7 +87,7 @@ int main() {
 
     // display values
     std::cout << "rescpu1 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << rescpu1[i] << " ";
     std::cout << "..." << std::endl << std::endl;
     
@@ -97,7 +97,7 @@ int main() {
 
     begin = clock();
     Eval<FUNCONVEXPF,CpuConv>::Run(Nx, Ny, f, oos2, x, y, b);
-    for(int i=0; i<vf.size(); i++)
+    for(size_t i=0; i<vf.size(); i++)
     	vf[i] = log(vf[i]);
     end = clock();
     std::cout << "time for CPU computation : " << double(end - begin) / CLOCKS_PER_SEC << std::endl;
@@ -107,13 +107,13 @@ int main() {
 
     // display values
     std::cout << "rescpu2 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << rescpu2[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
     // display mean of errors
     __TYPE__ s = 0;
-    for(int i=0; i<Nx*FUNCONVF::DIM; i++)
+    for(size_t i=0; i<Nx*FUNCONVF::DIM; i++)
         s += std::abs(rescpu1[i]-rescpu2[i]);
     std::cout << std::endl << "mean abs error = " << s/Nx << std::endl << std::endl;
 
@@ -129,7 +129,7 @@ int main() {
 
     // display values
     std::cout << "resgpu1 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << resgpu1[i] << " ";
     std::cout << "..." << std::endl << std::endl;
     
@@ -143,7 +143,7 @@ int main() {
 
     // display values
     std::cout << "resgpu2 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << resgpu2[i] << " ";
     std::cout << "..." << std::endl << std::endl;
     
@@ -167,7 +167,7 @@ int main() {
     rescpu1 = vf;
 
     std::cout << "rescpu1 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << rescpu1[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
@@ -181,7 +181,7 @@ int main() {
     resgpu1 = vf;
 
     std::cout << "resgpu1 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << resgpu1[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
@@ -197,7 +197,7 @@ int main() {
     resgpu2 = vf;
 
     std::cout << "resgpu2 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << resgpu2[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
@@ -205,19 +205,19 @@ int main() {
 	using GX2 = Grad<FUNCONVEXPF,X,E>;
     begin = clock();
 	Eval<GX2,GpuConv1D_FromHost>::Run(Nx, Ny, f, oos2, x, y, b, e);
-    for(int i=0; i<vf.size(); i++)
+    for(size_t i=0; i<vf.size(); i++)
     	vf[i] = vf[i]/exp(rescpu2[i/GX::DIM]);
     end = clock();
     std::cout << "time for GPU computation (1D scheme) : " << double(end - begin) / CLOCKS_PER_SEC << std::endl;
 
     resgpu2 = vf;
     std::cout << "rescpu2 = ";
-    for(int i=0; i<5; i++)
+    for(size_t i=0; i<5; i++)
         std::cout << resgpu2[i] << " ";
     std::cout << "..." << std::endl << std::endl;
 
     // display mean of errors
-    for(int i=0; i<Nx*GX::DIM; i++)
+    for(size_t i=0; i<Nx*GX::DIM; i++)
         s += std::abs(rescpu1[i]-resgpu2[i]);
     std::cout << std::endl << "mean abs error = " << s/Nx << std::endl << std::endl;
